requirements.txt
absl-py==0.15.0
argon2-cffi==21.1.0
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.10.0
dask==2021.9.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==2.2.6
django-crispy-forms==1.11.2
django-formtools==2.3
django-globals==0.3.2
djangorestframework==3.10.3
djangorestframework-simplejwt==4.6.0
entrypoints==0.3
flatbuffers==2.0
fsspec==2021.10.1
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
gunicorn==20.1.0
h5py==3.6.0
httpie==2.4.0
idna==2.10
imageio==2.9.0
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
keras-nightly==2.8.0.dev2021113008
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
locket==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.6.5
notebook==6.4.5
numpy==1.21.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
partd==1.2.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2019.3
PyWavelets==1.1.1
pywinpty==1.1.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.10.21
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.3.0
symspellpy==6.5.2
tb-nightly==2.8.0a20211130
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
textgenrnn==2.0.0
tf-estimator-nightly==2.8.0.dev2021113009
tf-nightly-cpu==2.8.0.dev20211130
threadpoolctl==3.0.0
tifffile==2021.10.12
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.3.0
widgetsnbextension==3.5.1
wrapt==1.13.3
django-colorfield
django-merged-inlines
django-npm
npm
django-compressor
django-compressor-toolkit